Moving stuff to workspace_utils.ts to reorganize, debug flags, cleaning up old code.
